// util/types.ts
export type CookieCartType = {
  itemId: number;
  itemQuantity: number;
};

export type DatabaseItemsType = {
  itemId: number;
  itemShortName: string;
  itemType: string;
  itemName: string;
  itemSubheadline: string;
  itemPrice: number;
  itemStockQuantity: number;
  itemDescription: string;
  itemFlag: string;
};